
\name{convCounts2Abun}
\alias{convCounts2Abun}
\docType{package}
\title{
  Convert an OTU count table to an OTU abundance table
}
\description{
  Convert an OTU count table to an OTU abundance table}
\usage{
  convCounts2Abun(otucounttable)  
}

\arguments{
 \item{otucounttable}{An OTU count table. Its rows correspond to OTUs, while its columns correspond to observations.}
}

\value{
  a matrix of OTU abundance  
}

\source{
Yushu Shi,Liangliang Zhang,Christine B. Peterson, Kim-Anh Do, Robert R. Jenq \emph{Performance determinants of
unsupervised clustering methods for microbiome data}  \emph{Accepted by Microbiome}}
