\name{Martinez}
\alias{Martinez}
\docType{data}
\title{
Martinez data
}
\description{
Martinez dataset includes OTU table, tree and sample geographic information.
 }

\source{
Martinez, I., Stegen, J. C., Maldonado-Gomez, M. X., Eren, A. M., Siba, P. M., Green-
hill, A. R., and Walter, J. (2015). \emph{The gut microbiota of rural Papua New Guineans:
composition, diversity patterns, and ecological processes.} Cell reports, 11(4):527-538.
}
