\name{Schnorr}
\alias{Schnorr}
\docType{data}
\title{
Schnorr data
}
\description{
Schnorr dataset includes OTU table, tree and sample geographic information.
 }

\source{

Schnorr, S. L., Candela, M., Rampelli, S., Centanni, M., Consolandi, C., Basaglia, G.,
Turroni, S., Biagi, E., Peano, C., Severgnini, M., et al. (2014). \emph{Gut microbiome of the
Hadza hunter-gatherers} Nature communications, 5:3654.
}
