\name{Smits}
\alias{Smits}
\docType{data}
\title{
Smits data
}
\description{
Smits dataset includes OTU table, tree and sample seasonal information. For the purpose of unsupervised clustering, we include only the most separated seasons, the early wet season and the late dry season.
 }

\source{

Smits, S. A., Leach, J., Sonnenburg, E. D., Gonzalez, C. G., Lichtman, J. S., Reid, G.,
Knight, R., Manjurano, A., Changalucha, J., Elias, J. E., Dominguez-Bello, M. G., and
Sonnenburg, J. L. (2017). \emph{Seasonal cycling in the gut microbiome of the hadza hunter-
gatherers of Tanzania.} Science, 357(6353):802-806.}
